CB-1837 Add RDS instance creation to cloud-aws
The cloud-aws module is augmented with most of the logic needed for
creating RDS instances. The core classes that have been modified or
added are:

* AwsResourceConnector - now delegates to the new AwsRdsLaunchService
  to create an RDS instance
* AwsRdsLaunchService - performs the RDS instance launch process:
  uses the following classes to create a CloudFormation template and
  creation request, then sends the request to AWS and polls for
  completion and outputs
* CloudFormationTemplateBuilder - builds the body of a CloudFormation
  template for a new RDS instance and DB subnet group from a context
* AwsStackRequestHelper - creates a CloudFormation stack creation
  request from a database stack, including setting the necessary
  template parameters

The actual CloudFormation template for new RDS instance, unlike the
one for new clusters, does not require much Freemarker processing at
this time, but that could change in the future.

Several views have been created or modified to codify interpreting the
general cloud-api model objects for RDS operations.

* AwsNetworkView - improved for better support for multiple subnets and
  avoidance of null pointer exceptions
* AwsRdsInstanceView - for locating RDS instance information
* AwsRdsDbSubnetGroupView - for locating DB subnet group information

Unit tests for much of the implementation work is included, but the
overall functionality has not yet been tested for real.